Reverse Proxy/ API Gateway — Step 03
```json
"host": "todo-api:8000"
"connection": "close"
```

- “host”: “todo-api:8000” — indicates the hostname and port number of the server to which the request is being sent (todo-api on port 8000). This information is exposed to the client because we forwarded the client requests directly to the todo-api server without any additional configurations.
“Revealing internal infrastructure details to the client is not ideal.”

In the next step of this project, we will change this behavior.
- “connection”: “close” — indicates that the connection will be closed after the current transaction is completed. This means that the client or server intends to close the TCP connection after sending or receiving the current HTTP message.
In an Nginx reverse proxy, this practice can help manage server resources by limiting the number of open connections. It is a part of the connection management strategy to control the lifecycle of connections between clients and servers, helping in the optimization of resource usage and performance improvement.