05-How can we login from the Next.js todo-client? - Add middleware.ts and logout button
In the current setup, to set the `_oauth2_proxy` cookie, we have to access http://localhost/.

However, in a real scenario, redirecting users out of your client application like this may not be ideal.

Middleware allows you to run code before a request is completed. Then, based on the incoming request, you can modify the response by rewriting, redirecting, modifying the request or response headers, or responding directly.